Switch to Ansible Collections structure
