+++ examples/hybrid/site.yml
@@ -2,6 +2,8 @@
 - hosts: localhost
   vars:
     ansible_python_interpreter: "./venv/bin/python"  # Required for virtualenv...
+  collections:
+    - pureport.pureport_ansible_modules
   roles:
     - role: aws
       aws_region: us-west-2
