--- meta/main.yml
@@ -1,29 +1,0 @@
----
-galaxy_info:
-  author: Matt Traynham
-  description: Ansible modules to support interaction with the Pureport ReST API
-  company: Pureport
-  license: GPLv3
-  min_ansible_version: 2.7
-  galaxy_tags:
-    - networking
-    - multicloud
-    - cloud
-    - aws
-    - azure
-    - google
-    - vpn
-    - pureport
-    - web
-    - sdn
-    - bgp
-    - nat
-  platforms:
-    - name: GenericLinux
-      versions:
-        - any
-  cloud_platforms:
-    - amazon
-    - azure
-    - google
-dependencies: []
