+++ plugins/modules/pureport_azure_express_route_connection.py
@@ -8,33 +8,21 @@ ANSIBLE_METADATA = {
 
 DOCUMENTATION = '''
 ---
-module: pureport_aws_direct_connect_connection
-short_description: Create, update or delete an AWS Direct Connect connection
+module: pureport_azure_express_route_connection
+short_description: Create, update or delete a Azure Express Route connection
 description:
-    - "Create, update or delete an AWS Direct Connect connection"
+    - "Create, update or delete a Azure Express Route connection"
 version_added: "2.8"
 requirements: [ pureport-client ]
 author: Matt Traynham (@mtraynham)
 options:
     network_href:
         required: true
-    aws_account_id:
+    service_key:
         description:
-            - The AWS account Id associated with the connection
+            - The Azure Express Route service key.
         required: true
         type: str
-    aws_region:
-        description:
-            - The AWS region associated with the connection
-        required: true
-        type: str
-    cloud_service_hrefs:
-        description:
-            - A list of cloud services for the connection
-            - This should be the full 'href' path to the CloudService ReST object (e.g /cloudServices/abc).
-        required: false
-        type: list
-        default: []
 extends_documentation_fragment:
     - pureport_client
     - pureport_network
@@ -46,23 +34,22 @@ extends_documentation_fragment:
 '''
 
 EXAMPLES = '''
-- name: Create a simple PRIVATE AWS Direct Connect connection for a network
-  pureport_aws_direct_connect_connection:
+- name: Create a simple PRIVATE Azure Express Route connection for a network
+  pureport_azure_express_route_connection:
     api_key: XXXXXXXXXXXXX
     api_secret: XXXXXXXXXXXXXXXXX
     network_href: /networks/network-XXXXXXXXXXXXXXXXXXXXXX
-    name: My Ansible AWS Direct Connect Connection
+    name: My Ansible Azure Express Route Connection
     speed: 50
     high_availability: true
     location_href: /locations/XX-XXX
     billing_term: HOURLY
-    aws_account_id: XXXXXXXXXXXX
-    aws_region: XX-XXXX-#
+    service_key: XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX
     wait_for_server: true  # Wait for the server to finish provisioning the connection
   register: result  # Registers the connection as the result
 
 - name: Update the newly created connection with changed properties
-  pureport_aws_direct_connect_connection:
+  pureport_azure_express_route_connection:
     api_key: XXXXXXXXXXXXX
     api_secret: XXXXXXXXXXXXXXXXX
     network_href: /networks/network-XXXXXXXXXXXXXXXXXXXXXX
@@ -71,13 +58,12 @@ EXAMPLES = '''
     high_availability: {{ result.high_availability }}
     location_href: {{ result.location.href }}
     billing_term: {{ result.billing_term }}
-    aws_account_id: {{ result.aws_account_id }}
-    aws_region: {{ result.aws_region }}
+    service_key: {{ result.service_key }}
     wait_for_server: true  # Wait for the server to finish updating the connection
   register: result  # Registers the connection as the result
 
 - name: Delete the newly created connection using the 'absent' state
-  pureport_aws_direct_connect_connection:
+  pureport_azure_express_route_connection:
     api_key: XXXXXXXXXXXXX
     api_secret: XXXXXXXXXXXXXXXXX
     network_href: /networks/network-XXXXXXXXXXXXXXXXXXXXXX
@@ -87,50 +73,44 @@ EXAMPLES = '''
     high_availability: {{ result.high_availability }}
     location_href: {{ result.location.href }}
     billing_term: {{ result.billing_term }}
-    aws_account_id: {{ result.aws_account_id }}
-    aws_region: {{ result.aws_region }}
+    service_key: {{ result.service_key }}
     wait_for_server: true  # Wait for the server to finish deleting the connection
 
-- name: Create a PRIVATE AWS Direct Connect connection with all properties configured
-  pureport_aws_direct_connect_connection:
+- name: Create a PRIVATE Azure Express Route connection with all properties configured
+  pureport_azure_express_route_connection:
     api_key: XXXXXXXXXXXXX
     api_secret: XXXXXXXXXXXXXXXXX
     network_href: /networks/network-XXXXXXXXXXXXXXXXXXXXXX
-    name: My Ansible AWS Direct Connect
+    name: My Ansible Azure Express Route Connection
     speed: 50
     high_availability: true
     location_href: /locations/XX-XXX
     billing_term: HOURLY
-    aws_account_id: XXXXXXXXXXXX
-    aws_region: XX-XXXX-#
+    service_key: XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX
     # Optional properties start here
-    description: My Ansible managed AWS Direct Connect connection
+    description: My Ansible managed Azure Express Route connection
     peering_type: PRIVATE
-    customer_asn: ######
     customer_networks:
       - address: a.b.c.d/x  # A valid CIDR address
-        name: My AWS accessible CIDR address
+        name: My Azure accessible CIDR address
     nat_enabled: true
     nat_mappings:
       - a.b.c.d/x  # A valid CIDR address, likely referencing a Customer Network
 
-- name: Create a PUBLIC AWS Direct Connect connection with all properties configured
-  pureport_aws_direct_connect_connection:
+- name: Create a PUBLIC Azure Direct Connect connection with all properties configured
+  pureport_azure_express_route_connection:
     api_key: XXXXXXXXXXXXX
     api_secret: XXXXXXXXXXXXXXXXX
     network_href: /networks/network-XXXXXXXXXXXXXXXXXXXXXX
-    name: My Ansible AWS Direct Connect
+    name: My Ansible Azure Express Route Connection
     speed: 50
     high_availability: true
     location_href: /locations/XX-XXX
     billing_term: HOURLY
-    aws_account_id: XXXXXXXXXXXX
-    aws_region: XX-XXXX-#
+    service_key: XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX
     # Optional properties start here
-    description: My Ansible managed AWS Direct Connect connection
+    description: My Ansible managed Azure Express Route connection
     peering_type: PUBLIC
-    cloud_service_hrefs:
-      - /cloudServices/aws-XX-XX-XXXX-#
 '''
 
 RETURN = '''
@@ -145,14 +125,14 @@ from ansible.module_utils.common.dict_transformations import \
     camel_dict_to_snake_dict, \
     snake_dict_to_camel_dict
 
-from ansible.module_utils.pureport.pureport import \
+from ansible_collections.pureport.pureport_ansible_modules.plugins.module_utils.pureport import \
     get_client_argument_spec, \
     get_client_mutually_exclusive, \
     get_network_argument_spec
-from ansible.module_utils.pureport.pureport_crud import \
+from ansible_collections.pureport.pureport_ansible_modules.plugins.module_utils.pureport_crud import \
     get_state_argument_spec, \
     get_resolve_existing_argument_spec
-from ansible.module_utils.pureport.pureport_connection_crud import \
+from ansible_collections.pureport.pureport_ansible_modules.plugins.module_utils.pureport_connection_crud import \
     get_wait_for_server_argument_spec, \
     get_connection_argument_spec, \
     get_cloud_connection_argument_spec, \
@@ -173,17 +153,13 @@ def construct_connection(module):
         'speed',
         'high_availability',
         'billing_term',
-        'customer_asn',
         'customer_networks',
-        'aws_account_id',
-        'aws_region'
+        'service_key'
     ))
     connection.update(dict(
-        type='AWS_DIRECT_CONNECT',
+        type='AZURE_EXPRESS_ROUTE',
         peering=dict(type=module.params.get('peering_type')),
         location=dict(href=module.params.get('location_href')),
-        cloud_services=[dict(href=cloud_service_href)
-                        for cloud_service_href in module.params.get('cloud_service_hrefs')],
         nat=dict(
             enabled=module.params.get('nat_enabled'),
             mappings=[dict(native_cidr=nat_mapping)
@@ -191,10 +167,6 @@ def construct_connection(module):
         )
     ))
     connection = snake_dict_to_camel_dict(connection)
-    # Correct naming
-    connection.update(dict(
-        customerASN=connection.pop('customerAsn')
-    ))
     return connection
 
 
@@ -210,9 +182,7 @@ def main():
     argument_spec.update(get_peering_connection_argument_spec())
     argument_spec.update(
         dict(
-            aws_account_id=dict(type='str', required=True),
-            aws_region=dict(type='str', required=True),
-            cloud_service_hrefs=dict(type='list', default=[])
+            service_key=dict(type='str', required=True)
         )
     )
     mutually_exclusive = []
