+++ plugins/modules/pureport_network_facts.py
@@ -8,66 +8,65 @@ ANSIBLE_METADATA = {
 
 DOCUMENTATION = '''
 ---
-module: pureport_cloud_region_facts
-short_description: Retrieve a list of cloud regions
+module: pureport_network_facts
+short_description: Retrieve a list of networks for an account
 description:
-    - "Retrieve a list of cloud regions"
+    - "Retrieve a list of networks for an account"
 version_added: "2.8"
 requirements: [ pureport-client ]
 author: Matt Traynham (@mtraynham)
+options:
+    account_href:
+        required: true
 extends_documentation_fragment:
     - pureport_client
+    - pureport_account
 '''
 
 EXAMPLES = '''
-- name: List cloud regions
-  pureport_cloud_region_facts:
+- name: List networks for an account
+  pureport_networks_facts:
     api_key: XXXXXXXXXXXXX
     api_secret: XXXXXXXXXXXXXXXXX
-  register: result   # Registers result.cloud_regions
+    account_href: /accounts/ac-XXXXXXXXXXXXXXXXXXXXXX
+  register: result   # Registers result.networks
 
-- name: Display all cloud region hrefs using a json_query filter
+- name: Display all network hrefs using a json_query filter
   debug:
     var: item
-  loop: "{{ result.cloud_regions | json_query('[*].href') }}"
+  loop: "{{ result.networks | json_query('[*].href') }}"
 '''
 
 RETURN = '''
-cloud_regions:
-    description: A list of CloudRegion (dict) objects.
+networks:
+    description: A list of Network (dict) objects.
     returned: success
     type: complex
     contains:
         id:
             description:
-                - The cloud region id.
+                - The network id.
             returned: success
             type: str
-            sample: "aws-us-west-1"
+            sample: "network-rfqj4qc9fO8hDOczEB7Z_Q"
         href:
             description:
-                - The cloud region href, a path to resource on the server.
+                - The network href, a path to resource on the server.
             returned: success
             type: str
-            sample: "/cloudRegions/aws-us-west-1"
-        provider:
+            sample: "/networks/network-rfqj4qc9fO8hDOczEB7Z_Q"
+        name:
             description:
-                - The cloud region provider.
+                - The name of the network.
             returned: success
             type: str
-            sample: "AWS"
-        provider_assigned_id:
+            sample: "My Network Name"
+        description:
             description:
-                - The cloud region provider's id.
+                - The description of the network.
             returned: success
             type: str
-            sample: "us-west-1"
-        display_name:
-            description:
-                - A display name for this cloud region.
-            returned: success
-            type: str
-            sample: "US West (N. California)"
+            sample: "My network description"
 '''
 
 from ansible.module_utils.basic import AnsibleModule
@@ -78,21 +77,24 @@ try:
     from pureport.exception.api import ClientHttpException
 except ImportError:
     ClientHttpException = None
-from ansible.module_utils.pureport.pureport import \
+from ansible_collections.pureport.pureport_ansible_modules.plugins.module_utils.pureport import \
     get_client_argument_spec, \
     get_client_mutually_exclusive, \
-    get_client
+    get_client, \
+    get_account_argument_spec, \
+    get_account
 
 
-def find_cloud_regions(module):
+def find_networks(module):
     """
-    List cloud regions
+    List networks
     :param AnsibleModule module: the ansible module
     """
     client = get_client(module)
+    account = get_account(module)
     try:
-        cloud_regions = client.cloud_regions.list()
-        module.exit_json(cloud_regions=[camel_dict_to_snake_dict(cloud_region) for cloud_region in cloud_regions])
+        networks = client.accounts.networks(account).list()
+        module.exit_json(networks=[camel_dict_to_snake_dict(network) for network in networks])
     except ClientHttpException as e:
         module.fail_json(msg=e.response.text, exception=format_exc())
 
@@ -100,13 +102,14 @@ def find_cloud_regions(module):
 def main():
     argument_spec = dict()
     argument_spec.update(get_client_argument_spec())
+    argument_spec.update(get_account_argument_spec(True))
     mutually_exclusive = []
     mutually_exclusive += get_client_mutually_exclusive()
     module = AnsibleModule(
         argument_spec=argument_spec,
         mutually_exclusive=mutually_exclusive
     )
-    find_cloud_regions(module)
+    find_networks(module)
 
 
 if __name__ == '__main__':
