+++ test/playbooks/ansible.cfg
@@ -1,4 +1,2 @@
 [defaults]
-library=../../library
-module_utils=../../module_utils
-doc_fragment_plugins=../../plugins/doc_fragments
+collections_paths=./collections
