+++ test/playbooks/pureport_port_connection.yml
@@ -1,5 +1,7 @@
 ---
 - hosts: localhost
+  collections:
+    - pureport.pureport_ansible_modules
   tasks:
     - name: Retrieve the access token for an api key and secret
       pureport_access_token_fact:
@@ -212,7 +214,7 @@
     - fail:
       when: result.changed != true
 
-    - name: Test update Non-HA Port connection (no changes)
+    - name: Test update HA Port connection (no changes)
       pureport_port_connection:
         api_base_url: "{{ api_base_url }}"
         api_access_token: "{{ access_token }}"
@@ -235,7 +237,7 @@
     - fail:
       when: result.changed == true
 
-    - name: Test update Non-HA Port connection (no id, no changes)
+    - name: Test update HA Port connection (no id, no changes)
       pureport_port_connection:
         api_base_url: "{{ api_base_url }}"
         api_access_token: "{{ access_token }}"
@@ -257,7 +259,7 @@
     - fail:
       when: result.changed == true
 
-    - name: Test update Non-HA Port connection (changes)
+    - name: Test update HA Port connection (changes)
       pureport_port_connection:
         api_base_url: "{{ api_base_url }}"
         api_access_token: "{{ access_token }}"
@@ -282,7 +284,7 @@
     - fail:
       when: result.changed != true
 
-    - name: Test update Non-HA Port connection (no id, changes)
+    - name: Test update HA Port connection (no id, changes)
       pureport_port_connection:
         api_base_url: "{{ api_base_url }}"
         api_access_token: "{{ access_token }}"
@@ -305,7 +307,7 @@
     - fail:
       when: result.changed != true
 
-    - name: Test delete Non-HA Port connection
+    - name: Test delete HA Port connection
       pureport_port_connection:
         api_base_url: "{{ api_base_url }}"
         api_access_token: "{{ access_token }}"
