+++ test/playbooks/pureport_site_ipsec_vpn_connection.yml
@@ -1,5 +1,7 @@
 ---
 - hosts: localhost
+  collections:
+    - pureport.pureport_ansible_modules
   tasks:
     - name: Retrieve the access token for an api key and secret
       pureport_access_token_fact:
